# openapi/Skoleoverblikket.Api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Skoleoverblikket API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/messages": {
      "post": {
        "tags": [
          "Messages"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesController_SendMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesController_SendMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesController_SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/{id}/read": {
      "post": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/messages/recipients": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessagesController_RecipientDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessagesController_RecipientDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessagesController_RecipientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notification-preferences": {
      "get": {
        "tags": [
          "NotificationPreferences"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationPreferencesController_NotificationPreferenceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationPreferencesController_NotificationPreferenceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationPreferencesController_NotificationPreferenceDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "NotificationPreferences"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NotificationPreferencesController_UpsertPreferenceItem"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NotificationPreferencesController_UpsertPreferenceItem"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NotificationPreferencesController_UpsertPreferenceItem"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessagesController_RecipientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/RecipientType"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MessagesController_SendMessageRequest": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "format": "uuid"
          },
          "recipientType": {
            "$ref": "#/components/schemas/RecipientType"
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NotificationPreferencesController_NotificationPreferenceDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "inApp": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NotificationPreferencesController_UpsertPreferenceItem": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "inApp": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NotificationType": {
        "enum": [
          "NewMessage",
          "NewContactMessage",
          "WeekPlanChanged",
          "AbsenceConfirmed",
          "AbsenceDismissed",
          "VacationRegistrationOpened"
        ],
        "type": "string"
      },
...
      "RecipientType": {
        "enum": [
          "Parent",
          "Staff",
          "Board"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}